Add files via upload
A newly established housing community association wants to develop a system to manage
the occupancy of their community. The system should be able to generate a block-wise list of
occupied and unoccupied houses. The details of the occupants should also be displayed based
on their flat numbers. The system should display the availability of houses based on the
BHKs for new clients and any new occupancy must be updated in the system. The system
should also track the payment of monthly maintenance charges for all the occupied flats.